+++ src/FilterTableHeader.cpp
@@ -22,14 +22,14 @@ FilterTableHeader::FilterTableHeader(QTableView* parent) :
     setContextMenuPolicy(Qt::CustomContextMenu);
 }
 
-void FilterTableHeader::generateFilters(int number, bool showFirst)
+void FilterTableHeader::generateFilters(size_t number, bool showFirst)
 {
     // Delete all the current filter widgets
     qDeleteAll(filterWidgets);
     filterWidgets.clear();
 
     // And generate a bunch of new ones
-    for(int i=0;i < number; ++i)
+    for(size_t i=0;i < number; ++i)
     {
         FilterLineEdit* l = new FilterLineEdit(this, &filterWidgets, i);
         if(!showFirst && i == 0)        // This hides the first input widget which belongs to the hidden rowid column
